feat: expose data processing via FastAPI
