+++ README.md
@@ -1,21 +1,25 @@
-# Projeto de Processamento de Dados de Autorizações Governamentais
+# API para Processamento de Dados de Autorizações Governamentais
 
-Este projeto foi desenvolvido para consolidar e normalizar dados sobre autorizações e provimentos governamentais a partir de diversos arquivos Excel. O objetivo é criar um dataset único e limpo, disponível nos formatos SQLite e CSV, para facilitar a análise e consulta.
+Este projeto foi desenvolvido para consolidar e normalizar dados sobre autorizações e provimentos governamentais a partir de diversos arquivos Excel. O objetivo é criar um dataset único e limpo, disponível através de uma API, e também nos formatos SQLite e CSV.
 
 ## Estrutura de Diretórios
 
 A organização do projeto segue a seguinte estrutura:
 
 ```
-.venv/
+.
+├── app/
+│   ├── __init__.py
+│   ├── main.py         # Lógica da API FastAPI
+│   └── processing.py   # Lógica de processamento dos dados
 ├── data/
-│   ├── raw/         # Contém os arquivos Excel brutos (.xlsx, .xls)
-│   └── processed/   # Contém os dados processados (autorizacoes.db e autorizacoes.csv)
+│   ├── raw/
+│   └── processed/
 ├── docs/
-│   └── dicionario.md # Documentação e dicionário de dados
 ├── scripts/
-│   └── process_data.py # Script principal para o processo de ETL
-└── requirements.txt    # Lista de dependências Python
+├── .gitignore
+├── README.md
+└── requirements.txt
 ```
 
 ## Requisitos e Instalação
@@ -34,23 +38,43 @@ Para executar este projeto, você precisará do Python 3 instalado.
     pip install -r requirements.txt
     ```
 
-## Como Usar
+## Como Executar a API
 
-Para processar os dados, siga os passos abaixo:
+Para iniciar a API, execute o seguinte comando a partir do diretório raiz do projeto:
+
+```sh
+uvicorn app.main:app --reload
+```
+
+A API estará disponível em `http://127.0.0.1:8000`.
+
+## Endpoints da API
+
+A API fornece os seguintes endpoints:
+
+*   `POST /process`: Inicia o processamento dos dados em background.
+*   `GET /data`: Retorna os dados processados do banco de dados com paginação.
+*   `GET /download`: Faz o download dos dados processados em um arquivo CSV.
+
+## Como Usar a API
 
 1.  **Adicione os Dados Brutos:**
     Certifique-se de que todos os arquivos Excel a serem processados estejam na pasta `data/raw/`.
 
-2.  **Execute o Script de Processamento:**
-    A partir do diretório raiz do projeto, execute o script principal:
+2.  **Inicie o Processamento:**
+    Envie uma requisição POST para o endpoint `/process` para iniciar o processamento dos dados.
     ```sh
-    python scripts/process_data.py
+    curl -X POST http://127.0.0.1:8000/process
     ```
-    O script irá ler todos os arquivos da pasta `data/raw`, consolidar os dados, normalizar os nomes das colunas, limpar as informações e salvar os resultados.
 
-## Saídas do Processamento
+3.  **Consulte os Dados:**
+    Após o processamento, você pode consultar os dados via o endpoint `/data`.
+    ```sh
+    curl http://127.0.0.1:8000/data?skip=0&limit=10
+    ```
 
-Após a execução bem-sucedida do script, os seguintes arquivos serão gerados na pasta `data/processed/`:
-
-*   `autorizacoes.db`: Um banco de dados SQLite contendo uma tabela chamada `autorizacoes` com todos os dados consolidados.
-*   `autorizacoes.csv`: Um arquivo CSV com os mesmos dados consolidados, ideal para importação em outras ferramentas de análise.
+4.  **Faça o Download dos Dados:**
+    Para baixar o dataset completo em CSV, acesse o endpoint `/download` em seu navegador ou use o `curl`:
+    ```sh
+    curl -o autorizacoes.csv http://127.0.0.1:8000/download
+    ```
