+++ requirements.txt
@@ -1,10 +1,24 @@
-et_xmlfile==2.0.0
+annotated-types==0.7.0
+anyio==4.11.0
+click==8.3.0
+et-xmlfile==2.0.0
+fastapi==0.118.0
+h11==0.16.0
+idna==3.10
 numpy==2.3.3
 openpyxl==3.1.5
 pandas==2.3.3
+pip==25.2
+pydantic==2.11.9
+pydantic-core==2.33.2
 python-dateutil==2.9.0.post0
 pytz==2025.2
 six==1.17.0
+sniffio==1.3.1
+starlette==0.48.0
+typing-extensions==4.15.0
+typing-inspection==0.4.2
 tzdata==2025.2
 uv==0.8.22
+uvicorn==0.37.0
 xlrd==2.0.2
